+++ packages/ui/src/alert-dialog/index.ts
@@ -1,3 +1,5 @@
+export type { PortalProps } from './portal';
 export { Root } from './root';
 export { Trigger } from './trigger';
 export { Popup } from './popup';
+export { Portal } from './portal';