+++ packages/ui/src/dialog/index.ts
@@ -3,8 +3,20 @@ import { CloseIcon } from './close-icon';
 import { Footer } from './footer';
 import { Header } from './header';
 import { Popup } from './popup';
+import { Portal } from './portal';
 import { Root } from './root';
 import { Title } from './title';
 import { Trigger } from './trigger';
 
-export { Action, CloseIcon, Footer, Header, Popup, Root, Title, Trigger };
+export type { PortalProps } from './portal';
+export {
+	Action,
+	CloseIcon,
+	Footer,
+	Header,
+	Popup,
+	Portal,
+	Root,
+	Title,
+	Trigger,
+};