+++ packages/ui/src/form/primitives/select/index.ts
@@ -1,4 +1,6 @@
 export { Item } from './item';
+export type { PortalProps } from './portal';
+export { Portal } from './portal';
 export { Popup } from './popup';
 export { Root } from './root';
 export { Trigger } from './trigger';