+++ packages/ui/src/form/primitives/select/popup.tsx
@@ -1,11 +1,12 @@
 import { Select as _Select } from '@base-ui/react/select';
 import clsx from 'clsx';
-import { forwardRef } from '@wordpress/element';
+import { cloneElement, forwardRef, isValidElement } from '@wordpress/element';
 import {
 	type ThemeProvider as ThemeProviderType,
 	privateApis as themePrivateApis,
 } from '@wordpress/theme';
 import { unlock } from '../../../lock-unlock';
+import { Portal } from './portal';
 import itemPopupStyles from '../../../utils/css/item-popup.module.css';
 import resetStyles from '../../../utils/css/resets.module.css';
 import styles from './style.module.css';
@@ -17,42 +18,47 @@ const ThemeProvider: typeof ThemeProviderType =
 
 export const Popup = forwardRef< HTMLDivElement, SelectPopupProps >(
 	function Popup(
-		{ className, container, children, style, ...restProps },
+		{ className, portal, children, style, ...restProps },
 		ref
 	) {
-		return (
-			<_Select.Portal container={ container }>
-				<_Select.Positioner
-					{ ...ITEM_POPUP_POSITIONER_PROPS }
-					alignItemWithTrigger={ false }
-					style={ style }
-					className={ clsx(
-						resetStyles[ 'box-sizing' ],
-						styles.positioner,
-						className
-					) }
-				>
-					<ThemeProvider>
-						<_Select.Popup
-							ref={ ref }
-							className={ itemPopupStyles.popup }
-							{ ...restProps }
-						>
-							<_Select.List className={ itemPopupStyles.list }>
-								<div
-									className={
-										itemPopupStyles[
-											'list-scrollable-container'
-										]
-									}
-								>
-									{ children }
-								</div>
-							</_Select.List>
-						</_Select.Popup>
-					</ThemeProvider>
-				</_Select.Positioner>
-			</_Select.Portal>
+		const rootPortal = portal ?? <Portal />;
+		const portalChildren = (
+			<_Select.Positioner
+				{ ...ITEM_POPUP_POSITIONER_PROPS }
+				alignItemWithTrigger={ false }
+				style={ style }
+				className={ clsx(
+					resetStyles[ 'box-sizing' ],
+					styles.positioner,
+					className
+				) }
+			>
+				<ThemeProvider>
+					<_Select.Popup
+						ref={ ref }
+						className={ itemPopupStyles.popup }
+						{ ...restProps }
+					>
+						<_Select.List className={ itemPopupStyles.list }>
+							<div
+								className={
+									itemPopupStyles[
+										'list-scrollable-container'
+									]
+								}
+							>
+								{ children }
+							</div>
+						</_Select.List>
+					</_Select.Popup>
+				</ThemeProvider>
+			</_Select.Positioner>
 		);
+
+		if ( isValidElement( rootPortal ) ) {
+			return cloneElement( rootPortal, { children: portalChildren } );
+		}
+
+		return <Portal>{ portalChildren }</Portal>;
 	}
 );