+++ packages/ui/src/tooltip/popup.tsx
@@ -1,12 +1,13 @@
 import clsx from 'clsx';
 import { Tooltip } from '@base-ui/react/tooltip';
-import { forwardRef } from '@wordpress/element';
+import { cloneElement, forwardRef, isValidElement } from '@wordpress/element';
 import {
 	type ThemeProvider as ThemeProviderType,
 	privateApis as themePrivateApis,
 } from '@wordpress/theme';
 import type { PopupProps } from './types';
 import { unlock } from '../lock-unlock';
+import { Portal } from './portal';
 import resetStyles from '../utils/css/resets.module.css';
 import styles from './style.module.css';
 
@@ -16,7 +17,7 @@ const ThemeProvider: typeof ThemeProviderType =
 const Popup = forwardRef< HTMLDivElement, PopupProps >( function TooltipPopup(
 	{
 		align = 'center',
-		container,
+		portal,
 		side = 'top',
 		sideOffset = 4,
 		children,
@@ -26,20 +27,20 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >( function TooltipPopup(
 	},
 	ref
 ) {
-	return (
-		<Tooltip.Portal container={ container }>
-			<Tooltip.Positioner
-				align={ align }
-				side={ side }
-				sideOffset={ sideOffset }
-				style={ style }
-				className={ clsx(
-					resetStyles[ 'box-sizing' ],
-					className,
-					styles.positioner
-				) }
-			>
-				{ /* This should ideally use whatever dark color makes sense,
+	const rootPortal = portal ?? <Portal />;
+	const portalChildren = (
+		<Tooltip.Positioner
+			align={ align }
+			side={ side }
+			sideOffset={ sideOffset }
+			style={ style }
+			className={ clsx(
+				resetStyles[ 'box-sizing' ],
+				className,
+				styles.positioner
+			) }
+		>
+			{ /* This should ideally use whatever dark color makes sense,
 				and not be hardcoded to #1e1e1e. The solutions would be to:
 				  - review the design of the tooltip, in case we want to stop
 				    hardcoding it to a dark background
@@ -48,18 +49,23 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >( function TooltipPopup(
 				    them;
 				  - remove the hardcoded `bg` setting from the `ThemeProvider` below
 					*/ }
-				<ThemeProvider color={ { bg: '#1e1e1e' } }>
-					<Tooltip.Popup
-						ref={ ref }
-						className={ styles.popup }
-						{ ...props }
-					>
-						{ children }
-					</Tooltip.Popup>
-				</ThemeProvider>
-			</Tooltip.Positioner>
-		</Tooltip.Portal>
+			<ThemeProvider color={ { bg: '#1e1e1e' } }>
+				<Tooltip.Popup
+					ref={ ref }
+					className={ styles.popup }
+					{ ...props }
+				>
+					{ children }
+				</Tooltip.Popup>
+			</ThemeProvider>
+		</Tooltip.Positioner>
 	);
+
+	if ( isValidElement( rootPortal ) ) {
+		return cloneElement( rootPortal, { children: portalChildren } );
+	}
+
+	return <Portal>{ portalChildren }</Portal>;
 } );
 
 export { Popup };